refactor: moved getTextColour outside component
