+++ client/src/components/Button.tsx
@@ -1,4 +1,5 @@
 import styles from "./Button.module.css";
+import { getTextColour } from "../utils/helpers";
 
 export interface ButtonProps {
   height?: string;
@@ -13,14 +14,6 @@ export interface ButtonProps {
   fontWeight?: string;
   onClick: () => void;
 }
-
-export const getTextColor = (hex: string) => {
-  const r = parseInt(hex.slice(1, 3), 16);
-  const g = parseInt(hex.slice(3, 5), 16);
-  const b = parseInt(hex.slice(5, 7), 16);
-  const yiq = (r * 299 + g * 587 + b * 114) / 1000;
-  return yiq >= 128 ? "#4F4F4F" : "#ffffff";
-};
 
 const Button = ({
   height,
@@ -49,7 +42,7 @@ const Button = ({
           border: backgroundColor,
           borderRadius: "10px",
           padding: "10px",
-          color: getTextColor(backgroundColor),
+          color: getTextColour(backgroundColor),
           fontFamily: "Montserrat",
           fontSize,
           fontWeight,
