+++ client/src/utils/helpers.tsx
@@ -1,0 +1,7 @@
+export const getTextColour = (hex: string) => {
+  const r = parseInt(hex.slice(1, 3), 16);
+  const g = parseInt(hex.slice(3, 5), 16);
+  const b = parseInt(hex.slice(5, 7), 16);
+  const yiq = (r * 299 + g * 587 + b * 114) / 1000;
+  return yiq >= 128 ? "#4F4F4F" : "#ffffff";
+};
